// src/interfaces/pet_interface.tsx
export interface Pet {
  bio: string;
  breed: string;
  color: string;
  date: string;
  dietary: string;
  height: number;
  height_category: string;
  hypoallergenic: boolean;
  name: string;
  owner: string;
  photo: string;
  status: string;
  type: string;
  weight: number;
  weight_category: string;
  __v: number;
  _id: string;
}